+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # RealmJoin Runbooks Changelog 
 
+## 2022-05-16
+* new runbook: Export all devices (to a storage account)
+
+## 2022-05-16
+* Many runbooks: Improve output 
+* device wipe and deveice outphase: Show owner/user UPN for the device
 ## 2022-05-12
 * Add User: Supports adding users to Exchange Distr. Lists / mail enabled groups